refactor: make ItemSerialNumValidator stateless by passing repository as parameter
